understanding and slightly refactoring frontend
