+++ frontend/src/hooks.server.js
@@ -24,7 +24,7 @@ export async function handle({ event, resolve }) {
     );
 
     if (!isPublicPage && !url.pathname.startsWith("/terminal") && !locals.user) {
-        throw redirect(303, `/csw/dashboard`);
+        throw redirect(303, `/dashboard`);
     }
 
     return resolve(event);
