+++ frontend/src/routes/admin/+page.svelte
